+++ src/auth/interfaces/jwt-payload-interface.ts
@@ -1,5 +1,9 @@
 // src/auth/jwt-payload.interface.ts
 export interface JwtPayload {
-  sub: string; // The identity UID
+  /**
+   * Optional namespace derived from `authContextResolver` or defaults to `root`
+   */
+  namespace?: string;     // Optional namespace derived from `authContextResolver` or defaults to `root`
+  sub: string;            // The identity UID
   sessionId: string;
 }