+++ src/auth/specs/auth.controller.spec.ts
@@ -6,12 +6,12 @@ import { AUTH_MODULE_OPTIONS } from '../auth.module';
 import { AuthTransport } from '../enums/auth-type.enum';
 import { ThrottlerGuard } from '@nestjs/throttler';
 import { Response, Request } from 'express';
-import { AuthCookieService } from '../core/cookie-namespace.resolver';
+import { AuthContextService } from '../core/auth-context.resolver';
 
 describe('AuthController', () => {
     let controller: AuthController;
     let authService: AuthService;
-    let cookieService: AuthCookieService;
+    let cookieService: AuthContextService;
 
     const mockOptions: AuthModuleOptions = {
         jwtSecret: 'test-secret',
@@ -41,7 +41,7 @@ describe('AuthController', () => {
             providers: [
                 { provide: AuthService, useValue: mockAuthService },
                 { provide: AUTH_MODULE_OPTIONS, useValue: mockOptions },
-                { provide: AuthCookieService, useValue: mockCookieService },
+                { provide: AuthContextService, useValue: mockCookieService },
             ],
         })
             .overrideGuard(ThrottlerGuard).useValue({ canActivate: () => true })
@@ -49,7 +49,7 @@ describe('AuthController', () => {
 
         controller = module.get<AuthController>(AuthController);
         authService = module.get<AuthService>(AuthService);
-        cookieService = module.get<AuthCookieService>(AuthCookieService);
+        cookieService = module.get<AuthContextService>(AuthContextService);
     });
 
     afterEach(() => {